server/requirements.txt
flask
flask_cors
python-dotenv
psycopg2
requests
bs4
gunicorn
python-dateutil
apscheduler